+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.11.0-dev
 
+- Added a special Archived state that is used to highlight tasks in the editor, but does not appear in the task lists. #37
+- Added support for custom keywords by keyword groupings for Active, Inactive, Waiting, Completed, and Archived tasks. #43
 - Added a new experimental features section in the settings.
 - Added experiemental support for extracting tasks from Org Mode files within the vault. #45